task_assignment/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.0
celery==5.3.6
redis==5.0.1
flask-jwt-extended==4.6.0
pydantic==2.5.2
python-dotenv==1.0.0
flask-limiter==3.5.0
